Translating BP into C++ and animation from C++ with AnimGraph nodes.
All blueprints have been translated into C++ classes except one, the
animation blueprint.
This one has been partially converted to keep code in phase with the UE
philosophy, ie animation are supposed to be controlled from the
animation graph.

General:
- More work is needed to polish/refactor, clean and document the code.

ILeapEventInterface:
- ILeapEventInterface can now be inherited from the game code, mostly
using the LEAPMOTION_API tag to export the code.
- ILeapEventInterface will call code overrides correctly. The BP and C++
interfaces need to be called individually with the current method.
Unfortunately the old method was only calling the BP function when
tested. Which makes sense since they were not declared as virtual due to
the BPcallable. It should still have worked, but it is the only
explanation I could think of that made sense.
- The interface in C++ is not implementing all the calls yet, only the
Execute_HandCountChanged and Execute_LeapHandMoved.

Animations:
- For now, the animation uses duplicates of the original blueprints.
They should be linked to the original ones later in the integration
process.
- The duplicates are inheriting from the LeapBaseAnim class. This new
class takes care of all the logic that was done in the blueprint event
graph (not the anim graph). Example tree: LeapBaseAnim <-
BasicCharacter_AnimBlueprint2 => This BP is used as a paramater in
LeapCharacter constructor to generate the animation blueprint.
This allows the blueprint to access the character, then using the
character's AnimBody through the AnimBody interface (inherited by the
character).
- Not all fingers are rigged yet. Blueprints are still in dev.

Editor:
- Adding a new editor module to the plugin to add nodes to the anim
graph.
- Integrating new animation nodes "Modify Hand" and "Modify Finger", to
stay within the UE guidelines.
- Modify Hand is rotating dynamicly the list of 15 bones specified in
the settings based on the rotation of the hand given as input.
- Modify Finger is rotating dynamicly the list of 3 bones specified in
the settings based on the 3 rotations given as input. (Old prototype,
soon deprecated)
- Nodes are still in dev.

Characters:
- Characters are now available in C++ too, meaning they can be directly
used in the game mode settings, or inherited.
- The basic rigged character and basic leap echo are available.